+++ BroadleafCommerce/src-framework/org/broadleafcommerce/rules/domain/ShoppingCartPromotion.java
@@ -1,57 +1,57 @@
 package org.broadleafcommerce.rules.domain;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 //@Entity
 //@Table(name = "BLC_PROMOTION_RULE_DEFINITION")
 public class ShoppingCartPromotion implements Serializable {
 
-	private static final long serialVersionUID = 1L;
+    private static final long serialVersionUID = 1L;
 
-//	@Id
-//	@GeneratedValue
-//	@Column(name = "SHOPPING_CART_PROMOTION_RULE_ID")
-	private Long id;
+    // @Id
+    // @GeneratedValue
+    // @Column(name = "SHOPPING_CART_PROMOTION_RULE_ID")
+    private Long id;
 
-//	@Column(name = "NAME")
-	private String name;
+    // @Column(name = "NAME")
+    private String name;
 
-//	@Column(name = "COUPON_CODE")
-	private String couponCode;
+    // @Column(name = "COUPON_CODE")
+    private String couponCode;
 
-//  @Column(name = "ORDER_TOTAL")
-	private double orderTotal;
+    // @Column(name = "ORDER_TOTAL")
+    private BigDecimal orderTotal;
 
-	public void setId(Long id) {
-		this.id = id;
-	}
+    public void setId(Long id) {
+        this.id = id;
+    }
 
-	public Long getId() {
-		return id;
-	}
+    public Long getId() {
+        return id;
+    }
 
-	public void setName(String name) {
-		this.name = name;
-	}
+    public void setName(String name) {
+        this.name = name;
+    }
 
-	public String getName() {
-		return name;
-	}
+    public String getName() {
+        return name;
+    }
 
-	public void setCouponCode(String couponCode) {
-		this.couponCode = couponCode;
-	}
+    public void setCouponCode(String couponCode) {
+        this.couponCode = couponCode;
+    }
 
-	public String getCouponCode() {
-		return couponCode;
-	}
+    public String getCouponCode() {
+        return couponCode;
+    }
 
-	public void setOrderTotal(double orderTotal) {
-		this.orderTotal = orderTotal;
-	}
+    public void setOrderTotal(BigDecimal orderTotal) {
+        this.orderTotal = orderTotal;
+    }
 
-	public double getOrderTotal() {
-		return orderTotal;
-	}
-
+    public BigDecimal getOrderTotal() {
+        return orderTotal;
+    }
 }
